move adapter class adapter class.
