+++ src/main/java/xyz/groundx/caver_ext_kas/rest_client/io/swagger/client/api/tokenhistory/model/AnyOfPageableTokenSummaryItems.java
@@ -13,6 +13,7 @@
 package xyz.groundx.caver_ext_kas.rest_client.io.swagger.client.api.tokenhistory.model;
 
 import com.google.gson.annotations.JsonAdapter;
+import xyz.groundx.caver_ext_kas.rest_client.io.swagger.client.api.tokenhistory.model.adapter.TokenSummaryAdapter;
 
 /**
 * AnyOfPageableTokenSummaryItemsItems
